refactor(iocuak-common): add modules
