+++ packages/iocuak-common/src/common/calculations/isFunction.ts
