+++ packages/iocuak-common/src/common/calculations/isPromiseLike.spec.ts
