+++ packages/iocuak-common/src/common/calculations/isPromiseLike.ts
