+++ packages/iocuak-common/src/containerModuleMetadata/models/ContainerModuleMetadataId.ts
@@ -1,3 +1,3 @@
-import { Newable } from './Newable';
+import { Newable } from '../../service/models/Newable';
 
 export type ContainerModuleMetadataId = string | symbol | Newable;
