+++ packages/iocuak-common/src/index.ts
@@ -1,12 +1,12 @@
-import { ContainerModuleMetadataId } from './models/ContainerModuleMetadataId';
-import { Newable } from './models/Newable';
-import { ServiceId } from './models/ServiceId';
-import { Tag } from './models/Tag';
-import { chain } from './utils/chain';
-import { isFunction } from './utils/isFunction';
-import { isPromiseLike } from './utils/isPromiseLike';
-import { mapIterable } from './utils/mapIterable';
-import { mapIterator } from './utils/mapIterator';
+import { mapIterator } from './common/actions/mapIterator';
+import { chain } from './common/calculations/chain';
+import { isFunction } from './common/calculations/isFunction';
+import { isPromiseLike } from './common/calculations/isPromiseLike';
+import { mapIterable } from './common/calculations/mapIterable';
+import { ContainerModuleMetadataId } from './containerModuleMetadata/models/ContainerModuleMetadataId';
+import { Newable } from './service/models/Newable';
+import { ServiceId } from './service/models/ServiceId';
+import { Tag } from './service/models/Tag';
 
 export type { ContainerModuleMetadataId, Newable, Tag, ServiceId };
